Fix UUID array issue and enable full application functionality
- Fixed database column type mismatch (Double to BigDecimal for hours fields)
- Added missing @Column mappings for RecurrenceSettings embedded class
- Created TaskServiceImpl to resolve dependency injection issues
- Implemented complete JWT authentication with Spring Security
- Added health check endpoints and OpenAPI/Swagger documentation
- Fixed Spring Security pattern parsing issue
- Created frontend test UI to verify backend connectivity
- Updated Tailwind CSS configuration for v4 compatibility
- All modules now fully functional with JPA enabled

The application runs successfully with:
- Backend on http://localhost:8080/api
- Frontend on http://localhost:3000
- Swagger UI on http://localhost:8080/api/swagger-ui/index.html

+++ taskava-backend/taskava-core-service/src/main/java/com/taskava/service/dto/CreateTaskRequest.java
@@ -8,6 +8,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
@@ -40,7 +41,7 @@ public class CreateTaskRequest {
     private LocalDate startDate;
     private LocalDate dueDate;
     
-    private Double estimatedHours;
+    private BigDecimal estimatedHours;
     private Integer storyPoints;
     
     private Set<UUID> followerIds;
